+++ src/naming-strategy/DefaultNamingStrategy.ts
@@ -1,14 +1,30 @@
 import {NamingStrategyInterface} from "./NamingStrategyInterface";
 import {RandomGenerator} from "../util/RandomGenerator";
 import {camelCase, snakeCase} from "../util/StringUtils";
+import {TableType} from "../metadata/types/TableTypes";
 
 /**
  * Naming strategy that is used by default.
  */
 export class DefaultNamingStrategy implements NamingStrategyInterface {
 
-    tableName(className: string, customName: string): string {
-        return customName ? customName : snakeCase(className);
+    /**
+     * Normalizes table name.
+     *
+     * @param targetName Name of the target entity that can be used to generate a table name.
+     * @param userSpecifiedName For example if user specified a table name in a decorator, e.g. @Entity("name")
+     */
+    tableName(targetName: string, userSpecifiedName: string): string {
+        return userSpecifiedName ? userSpecifiedName : snakeCase(targetName);
+    }
+
+    /**
+     * Creates a table name for a junction table of a closure table.
+     *
+     * @param originalClosureTableName Name of the closure table which owns this junction table.
+     */
+    closureJunctionTableName(originalClosureTableName: string): string {
+        return originalClosureTableName + "_closure";
     }
 
     columnName(propertyName: string, customName: string): string {
@@ -53,10 +69,6 @@ export class DefaultNamingStrategy implements NamingStrategyInterface {
         return camelCase(tableName + "_" + columnName);
     }
 
-    closureJunctionTableName(tableName: string): string {
-        return tableName + "_closure";
-    }
-
     foreignKeyName(tableName: string, columnNames: string[], referencedTableName: string, referencedColumnNames: string[]): string {
         const key = `${tableName}_${columnNames.join("_")}_${referencedTableName}_${referencedColumnNames.join("_")}`;
         return "fk_" + RandomGenerator.sha1(key).substr(0, 27); // todo: use crypto instead?
@@ -67,10 +79,13 @@ export class DefaultNamingStrategy implements NamingStrategyInterface {
     }
 
     /**
-     * Adds prefix to the table.
+     * Adds globally set prefix to the table name.
+     * This method is executed no matter if prefix was set or not.
+     * Table name is either user's given table name, either name generated from entity target.
+     * Note that table name comes here already normalized by #tableName method.
      */
-    prefixTableName(prefix: string, originalTableName: string): string {
-        return prefix + originalTableName;
+    prefixTableName(prefix: string|undefined, tableName: string): string {
+        return prefix ? prefix + tableName : tableName;
     }
 
 }