+++ src/schema-builder/schema/ColumnSchema.ts
@@ -56,7 +56,7 @@ export class ColumnSchema {
     constructor(options?: {
         name?: string,
         type?: string,
-        default?: string,
+        default?: any,
         isNullable?: boolean,
         isGenerated?: boolean,
         isPrimary?: boolean,